Implement ClubsPage component with club listing and detail views; integrate sidebar and enhance UI with animations and responsive design.
